api/requirements.txt
numpy
pillow
tensorflow-gpu
keras==2.3.1
Flask_cors 
twilio
prometheus_client
requests
firebase_admin
uwsgi
